// src/views/config/mail/utils/types.ts
// 虽然字段很少 但是抽离出来 后续有扩展字段需求就很方便了

interface FormItemProps {
  id: string;
  email: string;
  name: string;
  enabled: boolean;
  type: string;
  groupCode: string;
  priority: string;
  remark: string;
}
interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };